+++ synccommerce/src/sdk/connectionsadvanced.ts
@@ -3,6 +3,7 @@
  */
 
 import * as utils from "../internal/utils";
+import * as errors from "./models/errors";
 import * as operations from "./models/operations";
 import * as shared from "./models/shared";
 import { SDKConfiguration } from "./sdk";
@@ -11,7 +12,8 @@ import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
 /**
  * Create new and manage existing Sync for Commerce companies.
  */
-export class CompanyManagement {
+
+export class ConnectionsAdvanced {
     private sdkConfiguration: SDKConfiguration;
 
     constructor(sdkConfig: SDKConfiguration) {
@@ -19,10 +21,10 @@ export class CompanyManagement {
     }
 
     /**
-     * Create Sync for Commerce company
+     * Create company
      *
      * @remarks
-     * Creates a Codat company with a commerce partner data connection.
+     * Creates a Codat company..
      */
     async createCompany(
         req: shared.CreateCompany,
@@ -37,7 +39,7 @@ export class CompanyManagement {
             this.sdkConfiguration.serverURL,
             this.sdkConfiguration.serverDefaults
         );
-        const url: string = baseURL.replace(/\/$/, "") + "/meta/companies/sync";
+        const url: string = baseURL.replace(/\/$/, "") + "/companies";
 
         let [reqBodyHeaders, reqBody]: [object, any] = [{}, {}];
 
@@ -54,6 +56,7 @@ export class CompanyManagement {
 
         const headers = { ...reqBodyHeaders, ...config?.headers };
         headers["Accept"] = "application/json";
+
         headers[
             "user-agent"
         ] = `speakeasy-sdk/${this.sdkConfiguration.language} ${this.sdkConfiguration.sdkVersion} ${this.sdkConfiguration.genVersion} ${this.sdkConfiguration.openapiDocVersion}`;
@@ -94,6 +97,13 @@ export class CompanyManagement {
             case httpRes?.status == 200:
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.company = utils.objectToClass(JSON.parse(decodedRes), shared.Company);
+                } else {
+                    throw new errors.SDKError(
+                        "unknown content-type received: " + contentType,
+                        httpRes.status,
+                        decodedRes,
+                        httpRes
+                    );
                 }
                 break;
         }
@@ -105,7 +115,7 @@ export class CompanyManagement {
      * Create connection
      *
      * @remarks
-     * Create a data connection for company.
+     * Creates a connection for the company by providing a valid platformKey.
      */
     async createConnection(
         req: operations.CreateConnectionRequest,
@@ -120,11 +130,7 @@ export class CompanyManagement {
             this.sdkConfiguration.serverURL,
             this.sdkConfiguration.serverDefaults
         );
-        const url: string = utils.generateURL(
-            baseURL,
-            "/meta/companies/{companyId}/connections",
-            req
-        );
+        const url: string = utils.generateURL(baseURL, "/companies/{companyId}/connections", req);
 
         let [reqBodyHeaders, reqBody]: [object, any] = [{}, {}];
 
@@ -141,6 +147,7 @@ export class CompanyManagement {
 
         const headers = { ...reqBodyHeaders, ...config?.headers };
         headers["Accept"] = "application/json";
+
         headers[
             "user-agent"
         ] = `speakeasy-sdk/${this.sdkConfiguration.language} ${this.sdkConfiguration.sdkVersion} ${this.sdkConfiguration.genVersion} ${this.sdkConfiguration.openapiDocVersion}`;
@@ -181,6 +188,13 @@ export class CompanyManagement {
             case httpRes?.status == 200:
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.connection = utils.objectToClass(JSON.parse(decodedRes), shared.Connection);
+                } else {
+                    throw new errors.SDKError(
+                        "unknown content-type received: " + contentType,
+                        httpRes.status,
+                        decodedRes,
+                        httpRes
+                    );
                 }
                 break;
         }
@@ -189,83 +203,10 @@ export class CompanyManagement {
     }
 
     /**
-     * List companies
+     * List connections
      *
      * @remarks
-     * Retrieve a list of all companies the client has created.
-     */
-    async listCompanies(
-        req: operations.ListCompaniesRequest,
-        retries?: utils.RetryConfig,
-        config?: AxiosRequestConfig
-    ): Promise<operations.ListCompaniesResponse> {
-        if (!(req instanceof utils.SpeakeasyBase)) {
-            req = new operations.ListCompaniesRequest(req);
-        }
-
-        const baseURL: string = utils.templateUrl(
-            this.sdkConfiguration.serverURL,
-            this.sdkConfiguration.serverDefaults
-        );
-        const url: string = baseURL.replace(/\/$/, "") + "/meta/companies";
-
-        const client: AxiosInstance =
-            this.sdkConfiguration.securityClient || this.sdkConfiguration.defaultClient;
-
-        const headers = { ...config?.headers };
-        const queryParams: string = utils.serializeQueryParams(req);
-        headers["Accept"] = "application/json";
-        headers[
-            "user-agent"
-        ] = `speakeasy-sdk/${this.sdkConfiguration.language} ${this.sdkConfiguration.sdkVersion} ${this.sdkConfiguration.genVersion} ${this.sdkConfiguration.openapiDocVersion}`;
-
-        let retryConfig: any = retries;
-        if (!retryConfig) {
-            retryConfig = new utils.RetryConfig(
-                "backoff",
-                new utils.BackoffStrategy(500, 60000, 1.5, 3600000),
-                true
-            );
-        }
-        const httpRes: AxiosResponse = await utils.Retry(() => {
-            return client.request({
-                validateStatus: () => true,
-                url: url + queryParams,
-                method: "get",
-                headers: headers,
-                responseType: "arraybuffer",
-                ...config,
-            });
-        }, new utils.Retries(retryConfig, ["408", "429", "5XX"]));
-
-        const contentType: string = httpRes?.headers?.["content-type"] ?? "";
-
-        if (httpRes?.status == null) {
-            throw new Error(`status code not found in response: ${httpRes}`);
-        }
-
-        const res: operations.ListCompaniesResponse = new operations.ListCompaniesResponse({
-            statusCode: httpRes.status,
-            contentType: contentType,
-            rawResponse: httpRes,
-        });
-        const decodedRes = new TextDecoder().decode(httpRes?.data);
-        switch (true) {
-            case httpRes?.status == 200:
-                if (utils.matchContentType(contentType, `application/json`)) {
-                    res.companies = utils.objectToClass(JSON.parse(decodedRes), shared.Companies);
-                }
-                break;
-        }
-
-        return res;
-    }
-
-    /**
-     * List data connections
-     *
-     * @remarks
-     * Retrieve previously created data connections.
+     * List the connections for a company.
      */
     async listConnections(
         req: operations.ListConnectionsRequest,
@@ -280,11 +221,7 @@ export class CompanyManagement {
             this.sdkConfiguration.serverURL,
             this.sdkConfiguration.serverDefaults
         );
-        const url: string = utils.generateURL(
-            baseURL,
-            "/meta/companies/{companyId}/connections",
-            req
-        );
+        const url: string = utils.generateURL(baseURL, "/companies/{companyId}/connections", req);
 
         const client: AxiosInstance =
             this.sdkConfiguration.securityClient || this.sdkConfiguration.defaultClient;
@@ -292,6 +229,7 @@ export class CompanyManagement {
         const headers = { ...config?.headers };
         const queryParams: string = utils.serializeQueryParams(req);
         headers["Accept"] = "application/json";
+
         headers[
             "user-agent"
         ] = `speakeasy-sdk/${this.sdkConfiguration.language} ${this.sdkConfiguration.sdkVersion} ${this.sdkConfiguration.genVersion} ${this.sdkConfiguration.openapiDocVersion}`;
@@ -334,6 +272,13 @@ export class CompanyManagement {
                         JSON.parse(decodedRes),
                         shared.Connections
                     );
+                } else {
+                    throw new errors.SDKError(
+                        "unknown content-type received: " + contentType,
+                        httpRes.status,
+                        decodedRes,
+                        httpRes
+                    );
                 }
                 break;
         }
@@ -342,7 +287,7 @@ export class CompanyManagement {
     }
 
     /**
-     * Update data connection
+     * Update connection
      *
      * @remarks
      * Update a data connection
@@ -362,7 +307,7 @@ export class CompanyManagement {
         );
         const url: string = utils.generateURL(
             baseURL,
-            "/meta/companies/{companyId}/connections/{connectionId}",
+            "/companies/{companyId}/connections/{connectionId}",
             req
         );
 
@@ -381,6 +326,7 @@ export class CompanyManagement {
 
         const headers = { ...reqBodyHeaders, ...config?.headers };
         headers["Accept"] = "application/json";
+
         headers[
             "user-agent"
         ] = `speakeasy-sdk/${this.sdkConfiguration.language} ${this.sdkConfiguration.sdkVersion} ${this.sdkConfiguration.genVersion} ${this.sdkConfiguration.openapiDocVersion}`;
@@ -421,6 +367,13 @@ export class CompanyManagement {
             case httpRes?.status == 200:
                 if (utils.matchContentType(contentType, `application/json`)) {
                     res.connection = utils.objectToClass(JSON.parse(decodedRes), shared.Connection);
+                } else {
+                    throw new errors.SDKError(
+                        "unknown content-type received: " + contentType,
+                        httpRes.status,
+                        decodedRes,
+                        httpRes
+                    );
                 }
                 break;
         }