+++ synccommerce/src/sdk/sdk.ts
@@ -3,8 +3,9 @@
  */
 
 import * as utils from "../internal/utils";
-import { CompanyManagement } from "./companymanagement";
-import { Configuration } from "./configuration";
+import { ConfigurationAdvanced } from "./configurationadvanced";
+import { Connections } from "./connections";
+import { ConnectionsAdvanced } from "./connectionsadvanced";
 import { Integrations } from "./integrations";
 import * as shared from "./models/shared";
 import { Sync } from "./sync";
@@ -53,8 +54,8 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.1";
-    sdkVersion = "0.30.1";
-    genVersion = "2.58.0";
+    sdkVersion = "0.31.0";
+    genVersion = "2.84.3";
 
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);
@@ -62,26 +63,35 @@ export class SDKConfiguration {
 }
 
 /**
- * Sync for Commerce API: The API for Sync for Commerce. Sync for Commerce is an API and a set of supporting tools. It has been built to enable e-commerce, point of sale platforms to provide high-quality integrations with numerous accounting platform through standardized API, seamlessly transforming business sale's data into accounting artefacts.
+ * Sync for Commerce API: The API for Sync for Commerce.
  *
  * @remarks
- * [Read More...](https://docs.codat.io/sfc/overview)
+ *
+ * Sync for Commerce is an API and a set of supporting tools built to enable e-commerce and point of sale platforms to provide high-quality integrations with numerous accounting platform through standardized API, seamlessly transforming business sale's data into accounting artefacts.
+ *
+ * [Read More...](https://docs.codat.io/commerce/overview)
+ *
+ * <!-- Not seeing the end points you're expecting? We've reorganized our products, and you may be using a [different version of Sync for Commerce](https://docs.codat.io/sync-for-commerce-v1-api#/). -->
  */
 export class CodatSyncCommerce {
     /**
-     * Create new and manage existing Sync for Commerce companies.
-     */
-    public companyManagement: CompanyManagement;
-    /**
      * Expressively configure preferences for any given Sync for Commerce company.
      */
-    public configuration: Configuration;
+    public configurationAdvanced: ConfigurationAdvanced;
+    /**
+     * Create new and manage existing Sync for Commerce companies using the Sync flow UI.
+     */
+    public connections: Connections;
+    /**
+     * Create new and manage existing Sync for Commerce companies.
+     */
+    public connectionsAdvanced: ConnectionsAdvanced;
     /**
      * View useful information about codat's integrations.
      */
     public integrations: Integrations;
     /**
-     * Initiate a sync of Sync for Commerce company data into their respective accounting software.
+     * Initiate and monitor the sync of company data into accounting software.
      */
     public sync: Sync;
     /**
@@ -116,8 +126,9 @@ export class CodatSyncCommerce {
             serverURL: serverURL,
         });
 
-        this.companyManagement = new CompanyManagement(this.sdkConfiguration);
-        this.configuration = new Configuration(this.sdkConfiguration);
+        this.configurationAdvanced = new ConfigurationAdvanced(this.sdkConfiguration);
+        this.connections = new Connections(this.sdkConfiguration);
+        this.connectionsAdvanced = new ConnectionsAdvanced(this.sdkConfiguration);
         this.integrations = new Integrations(this.sdkConfiguration);
         this.sync = new Sync(this.sdkConfiguration);
         this.syncFlowPreferences = new SyncFlowPreferences(this.sdkConfiguration);