+++ synccommerce/RELEASES.md
@@ -511,3 +511,13 @@ Based on:
 - Speakeasy CLI 1.53.0 (2.58.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.30.1] https://www.npmjs.com/package/@codat/sync-for-commerce/v/0.30.1 - synccommerce
+
+## 2023-08-17 00:01:30
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.73.1 (2.84.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.31.0] synccommerce
+### Releases
+- [NPM v0.31.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/0.31.0 - synccommerce